backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pymongo==4.5.0
motor==3.3.1
python-dotenv==1.0.0
pydantic==2.0.2
requests==2.31.0
python-multipart==0.0.6
transformers==4.30.2
torch==2.0.1
prophet==1.1.4
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.0
scipy==1.10.1